常量引用、指针，代码优化、飞船模型挡风玻璃的拐角增加采样点

+++ SLDR/CFileReader.cpp
@@ -102,7 +102,7 @@ void CFileReader::ReadCURVE(LPCTSTR lpszPathName, vector<CP_PolyLine3D>* VT_Poly
 	int row = 0;
 	int numOfCurves = 0, degree = 3,numOfPoints=0;
 	sfile.ReadString(str);
-	numOfCurves = atof(str);
+	numOfCurves = atoi(str);
 	CP_PolyLine3D poly;
 	while (sfile.ReadString(str))
 	{
